+++ client.js
@@ -7,43 +7,14 @@ var React       = require('react');
 var ReactMount  = require('react/lib/ReactMount');
 var ReactAsync  = require('react-async');
 var Router      = require('react-router-component');
-var superagent  = require('superagent');
 
 var Pages       = Router.Pages;
 var Page        = Router.Page;
 
+var UserPage = require('./pages/UserPage.js');
+var MainPage = require('./pages/MainPage.js');
+
 ReactMount.allowFullPageRender = true;
-
-var MainPage = React.createClass({
-
-  render: function() {
-    return (
-      <div className="UserPage">
-        Hello, anonynous!
-        <a href="/users/andrey">Login</a>
-      </div>
-    );
-  }
-});
-
-var UserPage = ReactAsync.createClass({
-
-  getInitialStateAsync: function(cb) {
-    superagent.get(
-      'http://localhost:3000/api/users/' + this.props.username,
-      function(err, res) {
-        cb(err, res ? res.body : null);
-      });
-  },
-  
-  render: function() {
-    return (
-      <div className="MainPage">
-        {this.state.username ? 'Hello, ' + this.state.username : 'Loading...'}
-      </div>
-    );
-  }
-});
 
 var App = React.createClass({
 
@@ -53,7 +24,7 @@ var App = React.createClass({
       this.refs.router.navigate(e.target.attributes.href.value);
     }
   },
-  
+
   render: function() {
     return (
       <html>
